rename toggle to switch
