+++ src/css/astro.css
@@ -25,5 +25,5 @@
 @import url("components/status.css");
 @import url("components/tabs.css");
 @import url("components/toggle-legacy.css");
-@import url("components/toggle.css");
+@import url("components/switch.css");
 @import url("components/tree.css");
